+++ packages/tools/file/index.js
@@ -1,2 +1,3 @@
+export * from "./src/path";
 export * from "./src/read";
 export * from "./src/write";